+++ ocis-pkg/config/defaultconfig.go
@@ -1,6 +1,7 @@
 package config
 
 import (
+	"github.com/owncloud/ocis/v2/ocis-pkg/shared"
 	activitylog "github.com/owncloud/ocis/v2/services/activitylog/pkg/config/defaults"
 	antivirus "github.com/owncloud/ocis/v2/services/antivirus/pkg/config/defaults"
 	appProvider "github.com/owncloud/ocis/v2/services/app-provider/pkg/config/defaults"
@@ -37,7 +38,6 @@ import (
 	storageshares "github.com/owncloud/ocis/v2/services/storage-shares/pkg/config/defaults"
 	storageSystem "github.com/owncloud/ocis/v2/services/storage-system/pkg/config/defaults"
 	storageusers "github.com/owncloud/ocis/v2/services/storage-users/pkg/config/defaults"
-	store "github.com/owncloud/ocis/v2/services/store/pkg/config/defaults"
 	thumbnails "github.com/owncloud/ocis/v2/services/thumbnails/pkg/config/defaults"
 	userlog "github.com/owncloud/ocis/v2/services/userlog/pkg/config/defaults"
 	users "github.com/owncloud/ocis/v2/services/users/pkg/config/defaults"
@@ -52,6 +52,9 @@ func DefaultConfig() *Config {
 		Runtime: Runtime{
 			Port: "9250",
 			Host: "localhost",
+		},
+		Reva: &shared.Reva{
+			Address: "com.owncloud.api.gateway",
 		},
 
 		Activitylog:       activitylog.DefaultConfig(),
@@ -90,7 +93,6 @@ func DefaultConfig() *Config {
 		StorageShares:     storageshares.DefaultConfig(),
 		StorageSystem:     storageSystem.DefaultConfig(),
 		StorageUsers:      storageusers.DefaultConfig(),
-		Store:             store.DefaultConfig(),
 		Thumbnails:        thumbnails.DefaultConfig(),
 		Userlog:           userlog.DefaultConfig(),
 		Users:             users.DefaultConfig(),