+++ services/thumbnails/pkg/service/grpc/v0/service.go
@@ -15,7 +15,7 @@ import (
 	"github.com/cs3org/reva/v2/pkg/rgrpc/todo/pool"
 	"github.com/cs3org/reva/v2/pkg/storagespace"
 	"github.com/cs3org/reva/v2/pkg/utils"
-	"github.com/golang-jwt/jwt/v4"
+	"github.com/golang-jwt/jwt/v5"
 	"github.com/pkg/errors"
 	merrors "go-micro.dev/v4/errors"
 	"google.golang.org/grpc/metadata"