+++ services/thumbnails/pkg/service/jwt/jwt.go
@@ -1,6 +1,6 @@
 package jwt
 
-import "github.com/golang-jwt/jwt/v4"
+import "github.com/golang-jwt/jwt/v5"
 
 // ThumbnailClaims defines the claims for thumb-nailing
 type ThumbnailClaims struct {