+++ tests/acceptance/features/bootstrap/GraphContext.php
@@ -2840,6 +2840,30 @@ class GraphContext implements Context {
 	}
 
 	/**
+	 * @When /^user "([^"]*)" lists the activities for (?:folder|file) "([^"]*)" of space "([^"]*)" with (depth|limit) "([^"]*)" using the Graph API/
+	 *
+	 * @param string $user
+	 * @param string $resource
+	 * @param string $spaceName
+	 * @param string $filterType
+	 * @param string $filterValue
+	 *
+	 * @return void
+	 */
+	public function userListsTheActivitiesForFolderOfSpaceWithDepthOrLimitUsingTheGraphApi(string $user, string $resource, string $spaceName, string $filterType, string $filterValue): void {
+		$resourceId = $this->featureContext->spacesContext->getResourceId($user, $spaceName, $resource);
+		$response = GraphHelper::getActivities(
+			$this->featureContext->getBaseUrl(),
+			$this->featureContext->getStepLineRef(),
+			$user,
+			$this->featureContext->getPasswordForUser($user),
+			$resourceId,
+			[$filterType => $filterValue]
+		);
+		$this->featureContext->setResponse($response);
+	}
+
+	/**
 	 * @When the administrator gets federated users using the Graph API
 	 * @When user :user tries to get federated users using the Graph API
 	 *
@@ -2849,7 +2873,7 @@ class GraphContext implements Context {
 	 */
 	public function theUserGetsFederatedUsersUsingTheGraphApi(?string $user = null): void {
 		$credentials = $this->getAdminOrUserCredentials($user);
-	
+
 		$response = GraphHelper::getFederatedUsers(
 			$this->featureContext->getBaseUrl(),
 			$this->featureContext->getStepLineRef(),
@@ -2870,7 +2894,7 @@ class GraphContext implements Context {
 	 */
 	public function theUserGetsAllUsersUsingTheGraphApi(?string $user = null): void {
 		$credentials = $this->getAdminOrUserCredentials($user);
-	
+
 		$response = GraphHelper::getAllUsers(
 			$this->featureContext->getBaseUrl(),
 			$this->featureContext->getStepLineRef(),